requirements.txt
fastapi~=0.112.0
uvicorn==0.30.6
sqlalchemy==2.0
psycopg2-binary
python-dotenv~=1.0.1
python-multipart==0.0.9
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.9.0
